requirements.txt
python-telegram-bot==20.7
pymongo==4.6.0
python-dotenv==1.0.0
apscheduler==3.10.4
opencv-python==4.8.1.78
aiohttp==3.9.1
pydantic==2.5.2
motor==3.3.2
